requirements.txt
Brotli==1.0.9
cffi==1.17.1
contourpy==1.3.1
cupy-cuda12x==13.4.1
cycler==0.11.0
fastrlock==0.8.3
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
numpy==2.2.5
openrgb-python==0.3.3
packaging==24.2
pillow==11.1.0
pip==25.1
pycparser==2.22
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
setuptools==72.1.0
sip==6.10.0
six==1.17.0
sounddevice==0.5.2
tornado==6.5
wheel==0.45.1
